requirements.txt
Flask_Caching==1.10.1
Flask_Cors==3.0.10
Flask_SQLAlchemy==2.5.1
oss2==2.15.0
SQLAlchemy==1.4.31
